:recycle: Navigation Bar Constraint 업데이트 및 상속 클래스 생성
