+++ PersonalScheduler/Sources/Presentation/ScheduleDetail/View/Subviews/ScheduleDetailTitleView.swift
@@ -1,0 +1,10 @@
+//
+//  ScheduleDetailTitleView.swift
+//  PersonalScheduler
+//
+//  Copyright (c) 2023 Minii All rights reserved.
+        
+import UIKit
+
+final class ScheduleDetailTitleView: NavigationBar {
+}
